Use Arc/Mutex to deal with requirement for immutable handler.
Details: in order to make this damn application work, it was necessary
for me to use a combination of arc and mutex on the handler struct so
that it could be safely shared between multiple threads. Of course, it
can't actually handle concurrent requests, so what will actually happen
is it'll spit out an error saying "Can't handle request" (or something
along those lines) if it receives more than one request at a time, but
the idea is for this thing to provide data to a single client, so I
don't see that as a major risk.

Of course, I have yet to build the client.